+++ src/Pages/TodoList/TodoList.tsx
@@ -1,6 +1,7 @@
-import React, { useState, useRef } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import Header from 'Components/Header';
 import { Todocreate, TodoItem } from 'Components/Todo';
+import { useDragAndDrop } from 'Hooks/useDragAndDrop';
 import { useTodo } from 'Hooks/useTodo';
 import { style } from './TodoListStyle';
 import { Itodo, Status } from 'types';
@@ -16,42 +17,10 @@ const TodoList: React.FC = () => {
     selectStatusTodo,
   } = useTodo();
 
+  const { handleOnDragStart, handleOnDragOver, handleOnDragEnd } =
+    useDragAndDrop({ todoState, setTodoState });
+
   const [list, setList] = useState<Itodo[] | null>(null);
-  const dragStartIndex = useRef<number>(0);
-  const dragOverIndex = useRef<number>(0);
-  const handleOnDragStart = (
-    e: React.DragEvent<HTMLDivElement>,
-    index: number,
-  ): void => {
-    (e.target as HTMLDivElement).classList.add('dragging');
-    dragStartIndex.current = index;
-  };
-
-  const handleOnDragOver = (
-    e: React.DragEvent<HTMLDivElement>,
-    index: number,
-  ): void => {
-    dragOverIndex.current = index;
-  };
-
-  const handleOnDragEnd = (e: React.DragEvent<HTMLDivElement>): void => {
-    (e.target as HTMLElement).classList.remove('dragging');
-    const draggedEl = todoState.slice(
-      dragStartIndex.current,
-      dragStartIndex.current + 1,
-    );
-    const frontTodos = todoState.slice(0, dragOverIndex.current);
-    const rearTodos = todoState.slice(dragOverIndex.current, todoState.length);
-
-    if (dragOverIndex.current > dragStartIndex.current) {
-      frontTodos.splice(dragStartIndex.current, 1);
-    } else {
-      rearTodos.splice(dragStartIndex.current - dragOverIndex.current, 1);
-    }
-
-    const temp = [...frontTodos, ...draggedEl, ...rearTodos];
-    setTodoState(temp);
-  };
 
   return (
     <>
